backend/requirements.txt
fastapi
uvicorn

langchain
langchain-community
langchain-openai
langgraph

chromadb
pypdf
python-dotenv
openai
pydantic-settings
